telegram/telegram-channel-to-discord-message-forwarder/requirements.txt
PyYAML==6.0
Telethon==1.23.0
discord.py==1.7.3